// src/app/subModulos/interfaces/ubicacion/area-ubicacion-interface.ts
/*export interface MyHomeBrewery {
    taps: Array<Taps>;    
    barrels: Array<Barrels>;
}*/

export interface AreaInterface {

    iAreaId : string;
    cAreaDescripcion :string;
    iLocalId : string;
    cLocalDescripcion : string;
}

/*export interface Barrels {
    id : number;
    name: string;
    width: number;
    height: number;
    quantity: number;
}*/
  